Fix Transfer.Read when the readnow (want) is partially downloaded.
It is normal if the missing range starts to be split between have & want, but
this triggered a "simpleFetcher admitted failure" error before. Now, only
trigger the error when a range is not yet in have, but also not anymore in
want, indicating fetcher failure.

Also fix a bug where FindUncovered() would return partially covered ranges.