Add VFS readdir caching. Refs #2.
The cache size is configurable using commandline flags and defaults to 1000
entries (path & peer combinations). On a network with 5 peers, this means 200
directories can be cached. By setting the cache size to 0, the cache is
disabled.

On readdir, we always fanout to all peers. If a peer did not respond within
the 100 ms deadline, we try to pull its response from cache. If it comes in
later, we add it to the cache anyway so we respond correctly next time.

On open/stat, we try to pull responses from cache if they are recent enough;
this means that a normal `ls` only readdirs all peers once. (If they did not
respond in time to the initial readdir, and did not respond by the time of
the stat, they have no response in cache so they will be queried again.)