visualize: Improve accuracy of the transfer rates
Keep track of when a transfer saw its first and last traffic.
By keeping track of its first traffic, we can immediately show the correct speed rather than slowly ramping up to the correct number over 30s. This also smooths out whether or not the oldest bucket was already cleared, because we now have a more accurate denominator.
By keeping track of its last traffic, we can set it to zero if no new traffic was moved. If it was only a temporary dip, it'll show an average again as soon as a new metric comes in.