Fix udp test.
The current code connected to 0.0.0.0:n. This worked, but the reply was
addressed to 127.0.0.1:n, so the reply and resulting connection was considered
a separate stream by the multiplexer. Because of this, the peer initially
setting up the connection never got its reply, and sent INIT retransmissions
ending up at an already-established stream.