udptransport: Clean up timers we started for IO deadlines
If a client sets the deadline to be very high, each call to Read/Write would add a timer that'd stay alive until it fires. By calling t.Stop() we clean them up so they don't live past the Read/Write method.

We could write the code even better so we'd reuse the timer on future loop iterations if it wasn't changed, but we very rarely have a second iteration of the loop, and a third will roughly never happen. So I decided it wasn't worth the complexity.