Ensure UDP hole-punching is always used.
gRPC should keep a channel open to all endpoints, since UDP hole-punching
between two clients may only work if they are trying to connect to each other
via UDP. Hence, even if one client can connect to the other via TCP, and this
method of communication is preferred, the client should still also connect
using UDP; if not, the other client may not be able to connect to the first
client at all.

We get gRPC to work like this by using a balancer that connects to all
addresses, then using a picker that takes the best open connection and sends
all RPCs over it.