web: Use go:embed rather than turning everything into Go source code
This removes a 5MB go file, which is nice.

I decided to add the Git folder to remove the dependency on javascript
for CI / other contributors.

The dist folder is large, but separate files are slightly better than
one huge one because they might not all change and be huge blobs in Git.

I excluded the .js.map file which isn't needed for release builds. It
would get built into development binaries, which is actually nice.